extracted spawn capabilities
